CreditCardService htting WPAService using swift generics way
