+++ WPAssignment/WPAssignment/Utilities/WPAUIConstants.swift
@@ -1,0 +1,8 @@
+//
+//  WPAUIConstants.swift
+//  WPAssignment
+//
+//  Created by Karthick Mannarkudi Ramesh Kumar on 30/11/24.
+//
+
+import Foundation
