+++ WPAssignment/WPAssignment/Utilities/WPAURLConstants.swift
@@ -1,0 +1,14 @@
+//
+//  WPAURLConstants.swift
+//  WPAssignment
+//
+//  Created by Karthick Mannarkudi Ramesh Kumar on 30/11/24.
+//
+
+import Foundation
+
+class WPAURLConstants {
+    static let kBaseRandomAPIURL: String = "https://random-data-api.com"
+    static let kRandomAPIV2URL: String = "\(kBaseRandomAPIURL)/api/v2"
+}
+
